backend/requirements.txt
django==4.1.5
django-cors-headers==3.13.0  ## Para a nossa API, ou seja libera nossos dados para ser acessados
djangorestframework==3.14.0
django-environ==0.9.0 #A ideia deste pacote é unificar vários pacotes que fazem a mesma coisa
Pillow==9.4.0  #para trabalha com as imagens do Blog